+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## FUTURE
 
+* [ADDED] - `App.prototype.useSDK` method which integrates with `services` app `config` section
+* [ADDED] - `Route` constructor takes new optional `sdkMethodName` option
 * [REMOVED] - `connect-flash` peer dependency. The module was being explicitly loaded during session initialization `Route.prototype.useSession`
 * [FIXED] - implementation of handling the `baseUrl` config value. Routers are now aware of the root base path value. This way we don't break API for modules which depend on it.
 * [ADDED] - http json response fallback when request fails and the custom `error-response` event listener is registered